pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "global_health_ai_evals"
version = "0.1.0"
description = "Open Global Health & Biosecurity AI Evaluations — Domain 1: Cultural Contextual Validity"
readme = "README.md"
requires-python = ">=3.11"
license = "CC-BY-4.0"
authors = [
    { name = "Roki Seydi", email = "roki@holdinghealth.org" },
]
dependencies = [
    "inspect-ai>=0.3.190",
    "anthropic>=0.20.0",
    "openai>=1.0.0",
    "google-generativeai>=0.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "pandas>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "dataset_download: tests that require downloading external datasets",
]
